overhaul of field demos
- updating to reflect current best practices
