+++ src/main/java/com/webforj/samples/views/fields/textfield/TextFieldView.java
@@ -17,27 +17,28 @@ import com.webforj.router.annotation.Route;
 @FrameTitle("Text Field Form")
 public class TextFieldView extends Composite<FlexLayout> {
 
+  TextField nameField = new TextField();
+  TextField emailField = new TextField();
+  TextField telField = new TextField();
+  TextField urlField = new TextField();
+  TextField searchField = new TextField();
+  
   public TextFieldView() {
     getBoundComponent().setMargin("var(--dwc-space-l)").setSpacing("var(--dwc-space-l)")
     .setDirection(FlexDirection.ROW).setWrap(FlexWrap.WRAP);
         
-    TextField nameField = new TextField();
     nameField.setPlaceholder("Name").setType(TextField.Type.TEXT).setLabel("Enter Name")
     .setValue("John Doe");
         
-    TextField emailField = new TextField();
     emailField.setPlaceholder("Email").setType(TextField.Type.EMAIL).setLabel("Enter Email")
     .setValue("example@email.com");
         
-    TextField telField = new TextField();
     telField.setPlaceholder("Phone Number").setType(TextField.Type.TEL).setLabel("Enter Phone Number")
     .setValue("(123) 456-7890");
         
-    TextField urlField = new TextField();
     urlField.setPlaceholder("URL").setType(TextField.Type.URL).setLabel("Enter URL")
     .setValue("https://www.example.com");
-        
-    TextField searchField = new TextField();
+       
     searchField.setPlaceholder("Search").setType(TextField.Type.SEARCH).setLabel("Enter Your Search")
     .setValue("Search...");
