+++ src/main/resources/css/fields/colorfield/colorFieldDemo.css
@@ -1,22 +1,10 @@
-.main__window{
-  display: flex;
-  flex-direction: column;
-  margin: 20px;
-}
-
-.color__field{
-  width: 200px;
-  margin-bottom: 50px;
-}
-
 .display__panel{
   height: fit-content;
   width: fit-content;
   padding: 5px;
-  border: 1px solid black !important;
 }
 
-.color__panel{
+.colorDiv{
   background-color: black;
   margin-top: 5px;
   height: 100px;
