--- src/main/resources/css/fields/datefield/dateFieldDemo.css
@@ -1,10 +1,0 @@
-.main__window {
-  display: flex;
-  margin: 20px;
-  gap: var(--dwc-space-l);
-}
-
-.date__input {
-  flex: none;
-  width: 200px;
-}
