--- src/main/resources/css/fields/timefield/timeFieldMinMaxView.css
@@ -1,8 +1,0 @@
-.window {
-  display: flex;
-  align-items: flex-end;
-  margin: 20px;
-  gap: 50px;
-  width: 100%;
-  flex-wrap: wrap;
-}
